restructure context for wallets
