+++ deployables/app/app/components/wallet/ConnectWallet.tsx
@@ -1,29 +1,10 @@
 import { invariant } from '@epic-web/invariant'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { ZERO_ADDRESS } from '@zodiac/chains'
 import { type HexAddress, ProviderType } from '@zodiac/schema'
-import { getDefaultConfig } from 'connectkit'
 import { type ChainId } from 'ser-kit'
-import { createConfig, WagmiProvider } from 'wagmi'
-import { injected, metaMask, walletConnect } from 'wagmi/connectors'
+import { useDisconnect } from 'wagmi'
 import { Connect } from './Connect'
 import { Wallet } from './Wallet'
-
-const queryClient = new QueryClient()
-
-const WALLETCONNECT_PROJECT_ID = '0f8a5e2cf60430a26274b421418e8a27'
-
-const wagmiConfig = createConfig(
-  getDefaultConfig({
-    appName: 'Zodiac Pilot',
-    walletConnectProjectId: WALLETCONNECT_PROJECT_ID,
-    connectors: [
-      injected(),
-      metaMask(),
-      walletConnect({ projectId: WALLETCONNECT_PROJECT_ID }),
-    ],
-  }),
-)
 
 interface Props {
   pilotAddress: HexAddress | null
@@ -44,14 +25,10 @@ export const ConnectWallet = ({
   onConnect,
   onDisconnect,
 }: Props) => {
+  const { disconnect } = useDisconnect()
+
   if (pilotAddress == null || pilotAddress === ZERO_ADDRESS) {
-    return (
-      <QueryClientProvider client={queryClient}>
-        <WagmiProvider config={wagmiConfig}>
-          <Connect onConnect={onConnect} />
-        </WagmiProvider>
-      </QueryClientProvider>
-    )
+    return <Connect onConnect={onConnect} />
   }
 
   invariant(
@@ -60,15 +37,15 @@ export const ConnectWallet = ({
   )
 
   return (
-    <QueryClientProvider client={queryClient}>
-      <WagmiProvider config={wagmiConfig}>
-        <Wallet
-          chainId={chainId}
-          providerType={providerType}
-          pilotAddress={pilotAddress}
-          onDisconnect={onDisconnect}
-        />
-      </WagmiProvider>
-    </QueryClientProvider>
+    <Wallet
+      chainId={chainId}
+      providerType={providerType}
+      pilotAddress={pilotAddress}
+      onDisconnect={() => {
+        disconnect()
+
+        onDisconnect()
+      }}
+    />
   )
 }
