+++ deployables/app/app/components/wallet/index.ts
@@ -1,1 +1,2 @@
 export { ConnectWallet } from './ConnectWallet'
+export { WalletProvider } from './WalletProvider'
